changing datasource to Oracle Database 11g Enterprise Edition Release 11.2.0.4.0

+++ src/main/java/ua/com/integrity/dto/ParsedFiles.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ParsedFiles {
-    int auditId;
+    long auditId;
     String fileName;
 }
